Разбивает код джава скрипт на модули.
